Exercises: Flatten
