+++ code/README.md
@@ -4,12 +4,8 @@ Exercises
 **Installation**:
 `npm install`
 
-**Running tests**:
-Tests are located in their corresponding folders.  To run:
-
-```
-cd exercises/curry
-mocha *spec.js
-```
+**Running**:
+`npm run exercise`
+  where exercise = curry | compose | functors | applicative | monads
 
 Some will fail and some will pass. You'll need to edit the exercises until the tests pass.
