+++ code/monads/lib/eval_spec.js
@@ -1,0 +1,7 @@
+const
+  Spec = require('./specification'),
+  A    = require('../answers'),
+  E    = require('../exercises');
+
+Spec('Exercises',E);
+Spec('Answers'  ,A);
