--- code/part1/answers/compose/compose_exercises_spec.js
@@ -1,5 +1,0 @@
-const
-  E    = require('./compose_exercises'),
-  Spec = require('../../../specs/compose_exercises_spec');
-
-Spec(E);
