--- code/part1/exercises/curry/curry_exercises_spec.js
@@ -1,5 +1,0 @@
-const
-  E    = require('./curry_exercises'),
-  Spec = require('../../../specs/curry_exercises_spec');
-
-Spec(E);
