--- code/part2/answers/monads/monad_exercises_spec.js
@@ -1,5 +1,0 @@
-const
-  E    = require('./monad_exercises'),
-  Spec = require('../../../specs/monad_exercises_spec');
-
-Spec(E);
