--- code/part2/exercises/applicative/applicative_exercises_spec.js
@@ -1,5 +1,0 @@
-const
-  E    = require('./applicative_exercises'),
-  Spec = require('../../../specs/applicative_exercises_spec');
-
-Spec(E);
