--- code/part2/exercises/functors/functor_exercises_spec.js
@@ -1,5 +1,0 @@
-const
-  E    = require('./functor_exercises'),
-  Spec = require('../../../specs/functor_exercises_spec');
-
-Spec(E);
