+++ feedback_loop.md
@@ -3,25 +3,22 @@
 Clone the repo to a `local_folder`
 
 ```shell
-    git clone https://github.com/DrBoolean/mostly-adequate-guide.git    
+  git clone https://github.com/DrBoolean/mostly-adequate-guide.git
 ```
 
-change directory to the code folder in repository root 
+change directory to the code folder in repository root
 ```shell
-    cd /path/to/local_folder/mostly-adequate-guide/code/part1_exercises/
+  cd /path/to/local_folder/mostly-adequate-guide/code/
 ```
 
 install dependencies
 ```shell
-    npm install
-    # this picks up the dependencies from package.json and installs them
+  npm install
 ```
 
-run the test runner with the spec file for that exercise
-
+run the test runner with the spec file for that exercise, to see whats available
 ```shell
-    # for example
-    mocha exercises/curry/curry_exercises_spec.js -w
+  npm run
 ```
 
 this brings up the mocha test runner which shows the error immediately when you save the file with changed code. Achieve the requested refactor while keeping all the arrows green.
